requirements.txt
tensorflow
scikit-learn
tqdm
jax
pytorch-lightning
optax
numpy
einops
orbax-checkpoint
wandb